sidenav refactor
