+++ online-shopping-website/frontend/src/App.test.js
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import App from './App';
 import { ProductGrid } from './Components/Products/ProductGrid';
-import { SideNav } from './Components/SideNav';
+import { SideNav } from './Components/Products/SideNav';
 import Brands from './TestValues.json';
 import Sellers from './TestValues.json';
 import React from 'react';
