+++ online-shopping-website/frontend/src/Components/CartPage.js
@@ -1,21 +1,23 @@
-import Grid from '@mui/material/Grid';
 import * as React from 'react';
 import {useState} from 'react';
+import axios from 'axios';
+
+import Grid from '@mui/material/Grid';
+import Typography from "@mui/material/Typography";
 import Button from '@mui/material/Button';
 import {Link, useNavigate} from "react-router-dom";
 import Alert from '@mui/material/Alert';
-import CloseIcon from '@mui/icons-material/Close';
-import IconButton from '@mui/material/IconButton';
 import Collapse from '@mui/material/Collapse';
 import Stack from '@mui/material/Stack';
+import CloseIcon from '@mui/icons-material/Close';
+import IconButton from '@mui/material/IconButton';
 import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
 import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew';
 import TestData from '../TestValues.json';
-import axios from 'axios';
-import Typography from "@mui/material/Typography";
 
 export const CartPage = () => {
+    //Forces rerender of components on call
     const [state, setState] = useState(0);
     const forceUpdate = () => setState(state + 1);
 
@@ -29,17 +31,25 @@ export const CartPage = () => {
     const [alertVisible, setAlertVisible] = useState(false);
 
     const calculateCartTally = () => {
+        //calculating subtotal of all items
         cart.forEach((item) => {
             subtotal += ((item.quantity) * (item.price));
         });
+
+        //calculating taxes
         GST = subtotal * 0.05;
         QST = subtotal * 0.0975;
+
+        //calculating total after tax
         total = subtotal + GST + QST;
     }
 
     const PlaceOrder = () => {
+        //Opening order placement alert
         setAlertVisible(true);
         window.scrollTo(0, 0);
+
+        //Posting order to backend once place order has been clicked
         axios.post(process.env.REACT_APP_DB_CONNECTION + '/api/orders/create', {
             //TODO: replace with cookie values
             userId: 3,
@@ -52,10 +62,12 @@ export const CartPage = () => {
             console.log("Order addition to backend failed.");
         });
 
+        //Clearing cart
         cart = [];    //TODO: replace with delete cart cookie (on alert open)
         forceUpdate();
     }
 
+    //Displays cart items breakdown
     const PriceBreakdown = () => {
         return (
             cart.map((item, index) => {
@@ -74,9 +86,11 @@ export const CartPage = () => {
     }
 
     const CartItem = () => {
+        //instantiating navigation call
         let navigate = useNavigate();
 
         function IncrementItem(itemID) {
+            //Increment quantity, ensuring that quantity does not exceed maximum 10 items per product in the cart
             if (cart[itemID].quantity !== 10) {
                 cart[itemID].quantity++;
             }
@@ -84,6 +98,7 @@ export const CartPage = () => {
         }
 
         function DecreaseItem(itemID) {
+            //Decrement quantity, ensuring that quantity has at least 1 item per product in the cart
             if (cart[itemID].quantity !== 1) {
                 cart[itemID].quantity--;
             }
@@ -91,15 +106,19 @@ export const CartPage = () => {
         }
 
         function RemoveItem(itemID) {
+            //Creates new array containing every product in the cart except the one being removed
             let updatedCart = cart.filter((item) => item.id !== itemID);
 
+            //Reassigns all values of cart array except last index with updatedCart values
             for (let i = 0; i < updatedCart.length; i++) {
                 cart[i] = updatedCart[i];
             }
 
+            //Removes element at last index of cart
             cart.pop();
             forceUpdate();
 
+            //Show alert and navigate back to home when cart is empty
             if (cart.length === 0) {
                 window.alert("Cart emptied. Returning to home page.");
                 navigate(`/`);
@@ -110,14 +129,18 @@ export const CartPage = () => {
             cart.map((item, index) => {
                     return (
                         <Grid container className="CartItem" key={index}>
+                            {/*Item image*/}
                             <Grid item sm={1} md={2} sx={{position: 'relative'}}>
                                 <img className="CartItemImage" src={item.image} alt={item.name}/>
                             </Grid>
                             <Grid item xs={12} sm={10}>
                                 <Grid item xs={12} lg={12} sx={{display: 'flex'}}>
+                                    {/*Item Name*/}
                                     <Grid item sm={9} md={11}>
                                         <h3 style={{margin: '1rem 0'}}>{item.name}</h3>
                                     </Grid>
+
+                                    {/*Remove item button*/}
                                     <Grid item xs={3} md={1} sx={{margin: 'auto', textAlign: 'center'}}>
                                         <Button className="Cart-CloseButton"
                                                 onClick={() => RemoveItem(index)}>
@@ -127,18 +150,26 @@ export const CartPage = () => {
 
                                 </Grid>
                                 <Grid item xs={12} className='CartText'>
+
+                                    {/*Seller info*/}
                                     <Grid item xs={12} lg={3}>
                                         <h4 style={{margin: 0}}>Seller:</h4>
                                         <p style={{margin: '0.5rem 0'}}>{item.seller}</p>
                                     </Grid>
+
+                                    {/*Brand info*/}
                                     <Grid item xs={12} lg={3}>
                                         <h4 style={{margin: 0}}>Brand:</h4>
                                         <p style={{margin: '0.5rem 0'}}>{item.brand}</p>
                                     </Grid>
+
+                                    {/*Price info*/}
                                     <Grid item xs={12} lg={3}>
                                         <h4 style={{margin: 0}}>Price: {item.price} Ɖ</h4>
                                         <h4>Promotion: 20% off</h4>
                                     </Grid>
+
+                                    {/*Quantity buttons*/}
                                     <Grid item xs={12} lg={3}>
                                         <div>
                                             <h4 className="Cart-Quantity">Quantity</h4>
@@ -170,6 +201,8 @@ export const CartPage = () => {
     }
 
     calculateCartTally();
+
+    //When cart is empty diaplsy message
     if (cart == false) {
         return (
             <Grid container className="Cart-Container">
@@ -210,6 +243,8 @@ export const CartPage = () => {
             </Grid>
         )
     }
+
+    //When cart is full, display cart items
     return (
         <Grid container className="Cart-Container">
             <Collapse in={alertVisible} className="Cart-Alert">
